fqa,contry code, otp,email on referal not showing added

+++ Modules/FAQ/Routes/web.php
@@ -29,14 +29,21 @@ Route::group(['prefix' => 'app', 'as' => 'backend.', 'middleware' => ['auth','au
      * ---------------------------------------------------------------------
      */
 
-    Route::group(['prefix' => 'faqs', 'as' => 'faqs.'],function () {
+    Route::group(['prefix' => 'faqs', 'as' => 'faqs.', 'middleware' => ['permission:view_faqs']], function () {
       Route::get("index_list", [FAQController::class, 'index_list'])->name("index_list");
       Route::get("index_data", [FAQController::class, 'index_data'])->name("index_data");
-      Route::get('export', [FAQController::class, 'export'])->name('export');
-      Route::post('bulk-action', [FAQController::class, 'bulk_action'])->name('bulk_action');
-      Route::post('restore/{id}', [FAQController::class, 'restore'])->name('restore');
-      Route::delete('force-delete/{id}', [FAQController::class, 'forceDelete'])->name('force_delete');
-      Route::post('update-status/{id}', [FAQController::class, 'update_status'])->name('update_status');
+      Route::get('export', [FAQController::class, 'export'])->name('export')->middleware('permission:delete_faqs');
+      Route::post('bulk-action', [FAQController::class, 'bulk_action'])->name('bulk_action')->middleware('permission:delete_faqs');
+      Route::post('restore/{id}', [FAQController::class, 'restore'])->name('restore')->middleware('permission:delete_faqs');
+      Route::delete('force-delete/{id}', [FAQController::class, 'forceDelete'])->name('force_delete')->middleware('permission:delete_faqs');
+      Route::post('update-status/{id}', [FAQController::class, 'update_status'])->name('update_status')->middleware('permission:edit_faqs');
     });
-    Route::resource("faqs", FAQController::class);
+    Route::resource("faqs", FAQController::class)->middleware([
+        'index' => 'permission:view_faqs',
+        'create' => 'permission:add_faqs',
+        'store' => 'permission:add_faqs',
+        'edit' => 'permission:edit_faqs',
+        'update' => 'permission:edit_faqs',
+        'destroy' => 'permission:delete_faqs',
+    ]);
 });
